requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
anyio==3.6.2
apparmor==2.13.6
argilla==1.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.3.0
asn1crypto==1.4.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.2.0
Automat==22.10.0
Babel==2.8.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.9.3
black==23.1.0
bleach==6.0.0
blinker==1.4
blis==0.7.9
blobfile==2.0.1
Brlapi==0.8.4
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
comm==0.1.2
commonmark==0.9.1
confection==0.0.4
configobj==5.0.6
connexion==2.14.1
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.2
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distro==1.5.0
distro-info==1.0
dnspython==2.3.0
dropbox==11.36.0
email-validator==1.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
executing==1.2.0
Fabric3==1.14.post1
faiss-cpu==1.7.3
fastapi==0.93.0
fastjsonschema==2.16.2
filelock==3.9.0
flasgger==0.9.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==3.2.4
Flask-Limiter==3.1.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gpg===1.14.0-unknown
greenlet==1.1.3.post0
grpcio==1.48.1
grpcio-tools==1.48.1
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.18.1
httptools==0.5.0
httpx==0.23.3
hyperlink==21.0.0
idna==2.10
importlib-metadata==5.2.0
importlib-resources==5.12.0
incremental==22.10.0
inflection==0.5.1
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.12.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==8.0.3
jupyter-console==6.6.1
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.3.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langchain==0.0.105
langcodes==3.3.0
LibAppArmor==2.13.6
libmagic==1.0
limits==3.2.0
louis==3.23.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.28.1
mate-menu==20.4.1
mate-tweak==21.4.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mechanize==0.4.5
mistune==2.0.4
monotonic==1.6
msgpack==1.0.0
multidict==6.0.4
murmurhash==1.0.9
mutagen==1.45.1
mypy==0.991
mypy-extensions==0.4.3
mysqlclient==2.1.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.10.9
newsapi-python==0.2.7
nltk==3.8.1
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.23.5
nyx==2.1.0
oauthlib==3.2.2
olefile==0.46
onboard==1.4.1
onionshare==2.2
openai==0.26.1
openpyxl==3.1.1
ordered-set==4.1.0
orjson==3.8.7
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.11.0
parsel==1.7.0
parso==0.8.3
pathspec==0.11.0
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.13.0
ply==3.11
preshed==3.0.8
printful==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
Protego==0.2.1
protobuf==3.19.5
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.9.7
pycups==2.0.1
pycurl==7.43.0.6
pydantic==1.10.5
PyDispatcher==2.0.7
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyMySQL==1.0.2
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==23.0.0
pyparsing==2.4.7
pypdf2==3.0.1
PyPrintful==1.0.0
PyQt5==5.15.2
PyQt5-sip==12.10.1
pyrsistent==0.18.1
PySimpleSOAP==1.16.2
pysmbc==1.0.23
PySocks==1.7.1
python-apt==2.2.1+parrot1
python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
python-docx==0.8.11
python-dotenv==0.15.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-xapp==2.0.2
python-xlib==0.29
pytz==2021.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
regex==2022.10.31
reportbug==7.10.3+deb11u1
reportlab==3.6.12
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
schedule==1.1.0
scikit-learn==1.2.1
scipy==1.10.1
Scrapy==2.8.0
selectorlib==0.16.0
Send2Trash==1.8.0
service-identity==21.1.0
setproctitle==1.2.1
simplejson==3.17.2
six==1.16.0
sklearn==0.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.2.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.41
srsly==2.4.6
stability-sdk==0.3.1
stack-data==0.6.2
starlette==0.25.0
stem==1.8.0
stone==3.3.1
stripe==5.1.1
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.2.1
terminado==0.17.1
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.1.2
tinycss2==1.2.1
tldextract==3.4.0
tomli==2.0.1
torbrowser-launcher==0.3.3
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
Twisted==22.10.0
twitter==1.19.6
typer==0.7.0
typing-extensions==4.4.0
typing-inspect==0.8.0
ufw==0.36
ujson==5.7.0
unattended-upgrades==0.1
unstructured==0.5.2
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.20.0
uvloop==0.17.0
validators==0.19.0
virtualenv==20.17.1
w3lib==2.1.1
wasabi==1.1.1
watchfiles==0.18.1
wcwidth==0.2.6
weaviate-client @ git+https://github.com/semi-technologies/weaviate-python-client.git@3554e3237f8f6c80f7d9b353ad7b81be5b4e6c4b
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
WTForms==3.0.1
xdg==5
XlsxWriter==3.0.8
yarl==1.8.2
youtube-dl==2021.6.6
zipp==3.11.0
zope.interface==5.5.2
